src/test/java/com/algorithm/dataStructure/matrix/SpiralOrderMatrixTest.java
package com.algorithm.dataStructure.matrix;

import org.junit.Assert;
import org.junit.Test;

import java.util.Arrays;

public class SpiralOrderMatrixTest {

    SpiralOrderMatrix service = new SpiralOrderMatrix();

    @Test
    public void spiralOrder() {

        Assert.assertEquals(Arrays.asList(1, 2, 3, 6, 9, 8, 7, 4, 5), service.spiralOrder(new int[][]{{1, 2, 3}, {4, 5, 6}, {7, 8, 9}}));
        Assert.assertEquals(Arrays.asList(1, 2, 3, 4, 8, 12, 11, 10, 9, 5, 6, 7), service.spiralOrder(new int[][]{{1, 2, 3, 4}, {5, 6, 7, 8}, {9, 10, 11, 12}}));
    }

    @Test
    public void spiralOrderOfficial() {
        Assert.assertEquals(Arrays.asList(1, 2, 3, 6, 9, 8, 7, 4, 5), service.spiralOrderOfficial(new int[][]{{1, 2, 3}, {4, 5, 6}, {7, 8, 9}}));
        Assert.assertEquals(Arrays.asList(1, 2, 3, 4, 8, 12, 11, 10, 9, 5, 6, 7), service.spiralOrderOfficial(new int[][]{{1, 2, 3, 4}, {5, 6, 7, 8}, {9, 10, 11, 12}}));

    }
}